+++ pyuploadcare/api/api.py
@@ -7,9 +7,8 @@ from time import time
 from typing import Any, Dict, Iterable, List, Optional, Type, Union, cast
 from uuid import UUID
 
-from httpx._types import RequestFiles
-
 from pyuploadcare.api import entities, responses
+from pyuploadcare.api._httpx import RequestFiles
 from pyuploadcare.api.addon_entities import (
     AddonExecutionGeneralRequestData,
     AddonExecutionParams,